==> Lessons/10_GLMs.Rmd <==
---
title: "10: Generalized Linear Models (T-Test)"
author: "Environmental Data Analytics | Kateri Salk"
date: "Spring 2020"
output: pdf_document
geometry: margin=2.54cm
editor_options: 
  chunk_output_type: console
---

## Objectives
1. Describe the components of the generalized linear model (GLM)
2. Apply special cases of the GLM (t-test) to real datasets
3. Interpret and report the results of t-tests in publication-style formats

## Set up
```{r, message = FALSE}
getwd()
library(tidyverse)

EPAair <- read.csv("./Data/Processed/EPAair_O3_PM25_NC1819_Processed.csv") #can use _KS

# Set date to date format
EPAair$Date <- as.Date(EPAair$Date, format = "%Y-%m-%d")

# Set theme
mytheme <- theme_classic(base_size = 14) + #base size is larger than usual, why???
  theme(axis.text = element_text(color = "black"), 
        legend.position = "top")
theme_set(mytheme)
```

## Generalized Linear Models (GLMs)

The one-sample test (model of the mean), two-sample t-test, analysis of variance (ANOVA), and linear regression are all special cases of the **generalized linear model** (GLM). The GLM also includes analyses not covered in this class, including logistic regression, multinomial regression, chi square, and log-linear models. The common characteristic of general linear models is the expression of a continuous response variable as a linear combination of the effects of categorical or continuous explanatory variables, plus an error term that expresses the random error associated with the coefficients of all explanatory variables. The explanatory variables comprise the deterministic component of the model, and the error term comprises the stochastic component of the model. Historically, artificial distinctions were made between linear models that contained categorical and continuous explanatory variables, but this distinction is no longer made. The inclusion of these models within the umbrella of the GLM allows models to fit the main effects of both categorical and continuous explanatory variables as well as their interactions. 

### Choosing a model from your data: A "cheat sheet"

**T-test:** Continuous response, one categorical explanatory variable with two categories (or comparison to a single value if a one-sample test).

**One-way ANOVA (Analysis of Variance):** Continuous response, one categorical explanatory variable with more than two categories.

**Two-way ANOVA (Analysis of Variance)** Continuous response, two categorical explanatory variables.

**Single Linear Regression** Continuous response, one continuous explanatory variable.

**Multiple Linear Regression** Continuous response, two or more continuous explanatory variables.

**ANCOVA (Analysis of Covariance)** Continuous response, categorical explanatory variable(s) and  continuous explanatory variable(s).

If multiple explanatory variables are chosen, they may be analyzed with respect to their **main effects** on the model (i.e., their separate impacts on the variance explained) or with respsect to their **interaction effects,** the effect of interacting explanatory variables on the model. 

### Assumptions of the GLM

The GLM is based on the assumption that the data residuals approximate a normal distribution (or a linearly transformed normal distribution). We will discuss the non-parametric analogues to several of these tests if the assumptions of normality are violated. For tests that analyze categorical explanatory variables, the assumption is that the variance in the response variable is equal among groups. Note: environmental data often violate the assumptions of normality and equal variance, and we will often proceed with a GLM even if these assumptions are violated. In this situation, justifying the decision to proceed with a linear model must be made. 

## T-Test
### One-sample t-test
The object of a one sample test is to test the null hypothesis that the mean of the group is equal to a specific value. For example, we might ask ourselves (from the EPA air quality processed dataset): 

Are Ozone levels below the threshold for "good" AQI index (0-50)?

```{r}

summary(EPAair$Ozone) #test mean of group is equal to specifi value
EPAair.subsample <- sample_n(EPAair, 5000) #I'd like to take random assortment of 5000

# Evaluate assumption of normal distribution
shapiro.test((EPAair.subsample$Ozone)) #p value is really small, so not well apprix by normal distr is really low??? just stating p value means nothing without context
ggplot(EPAair, aes(x = Ozone)) +
  geom_histogram() #not normal, common problem
qqnorm(EPAair$Ozone); qqline(EPAair$Ozone) #semi colon asks you to continue line, theoretical quantiles and sample quantiles;

O3.onesample <- t.test(EPAair$Ozone, mu = 50, alternative = "less") #save it as an object so you can call back to it; alt less means alt hypothese means mean is less than 50, better than two tailed bc just a one sample t test; need alt so it doens't think you're looking at both greater than and less than; don't include Q plot in report
O3.onesample

Ozone.plot <- ggplot(EPAair, aes(x = Ozone)) +
  #geom_density(stat = "count", fill = "gray") + #can uncode this and # out the below line which gives you more jagged distribution bc it shows count of data
  geom_density(fill = "gray") +
  geom_vline(xintercept = 50, color = "#238b45", lty = 2, size = 0.9) + #testing against mu of 50, lty 2 is dashed line, size = thickness
  scale_x_continuous(expand = c(0, 0)) + scale_y_continuous(expand = c(0, 0)) #can make graph line up with zero, have to specify for both x and y axes; green line is null hypothesis
print(Ozone.plot) #only needs x aesthetic
```

Write a sentence or two about the results of this test. Include both the results of the test and an interpretation that puts the findings in context of the resarch question.

> Via the t-test, the majority of the results fall below the true mean of 50. no...
We can accept that the null hypothesis is less than 50. The t-test was less than _ and mean was less than confidence interval...
CORRECT SENTENCE: EPA ozone measurements for 2018-2019 were significantly lower than 50, the AQI threshold for "good" air quality (t-test; t = -57.98, p < 0.0001) #sometimes putting p value is not useful if it's really small, don't include null hypothesis, comm what test was to avg reader; goes in final report (results)

### Two-sample t-test
The two-sample *t* test is used to test the hypothesis that the mean of two samples is equivalent. Unlike the one-sample tests, a two-sample test requires a second assumption that the variance of the two groups is equivalent. Are Ozone levels different between 2018 and 2019?

```{r}
shapiro.test(EPAair$Ozone[EPAair$Year == 2018])
shapiro.test(EPAair$Ozone[EPAair$Year == 2019])
var.test(EPAair$Ozone ~ EPAair$Year) #were ozone levels diff btwn 18 and 19, subset with square brackets, not well approx by a normal distr bc very low p values so our null is that sample is well approx by normal distrib, tilda is "by"; null would be that it IS = to 1

ggplot(EPAair, aes(x = Ozone, color = as.factor(Year))) +
  geom_freqpoly() #freq polygon colored by year as.factor year bc you're counting the years 2018 and 2019 as two different values, can do it in middle of ggplot, don't mess with data frame but specify as a factor, fairly similar distribution

# Format as a t-test, test ozone by (tilda) year, alt hyp is that means are diff (default for t-test)
O3.twosample <- t.test(EPAair$Ozone ~ EPAair$Year)
O3.twosample #reject the null hypothesis that means are equal, t value is negative so we get p value less than ? 95% confid int gives us confi of? **** about a 1 unit increase in AQI from 18 to 19, over 6000 degress of freedom, higher the number the better
O3.twosample$p.value #can call up items from the lists, just a different formatting

# Format as a GLM, can put linear combo to predict response variable, can do this with lm function
O3.twosample2 <- lm(EPAair$Ozone ~ EPAair$Year)
summary(O3.twosample2) #summary is for linear model, for each inc in yr, we're seeing x increase in which value??, note rounding errors but same output more or less

plot(O3.twosample2) #use plot function for t dot test you can't plot this way bc can only call lm ; shows fit vs how off it was from each fit (residuals); hit return again (looks pretty much the same from earlier) using residuals instead of ?? looking for what fall's out of cook's distance = outlier in data set

plot(O3.twosample2)
par(mfrow = c(2,2)) #any time I plot plot 4 at once in 2 rows in 2 columns; have to run above line with it first to get all 4 at once

```

### Non-parametric equivalent of t-test: Wilcoxon test

When we wish to avoid the assumption of normality, we can apply *distribution-free*, or non-parametric, methods in the form of the Wilcoxon rank sum (Mann-Whitney) test. The Wilcoxon test replaces the data by their rank and calculates the sum of the ranks for each group. Notice that the output of the Wilcoxon test is more limited than its parametric equivalent.

```{r}
#just looking at raw data, takes data and ranks form lowest to highest and then comparing sum of ranks within each group

O3.onesample.wilcox <- wilcox.test(EPAair$Ozone, mu = 50, alternative = "less")
O3.onesample.wilcox #gives v and p value, shows our data are significantly less than 50
O3.twosample.wilcox <- wilcox.test(EPAair$Ozone ~ EPAair$Year)
O3.twosample.wilcox #gives you a w instead of a V - what does this mean???*****
```

### Visualization and interpretation challenge

Create three plots, each with appropriately formatted axes and legends. Choose a non-default color palette.

1. geom_density of ozone divided by year (distinguish between years by adding transparency to the geom_density layer).
2. geom_boxplot of ozone divided by year . Add letters representing a significant difference between 2018 and 2019 (hint: stat_summary). 
3. geom_violin of ozone divided by year, with the 0.5 quantile marked as a horizontal line. Add letters representing a significant difference between 2018 and 2019. 

```{r}
#1
Geomdens.plot <- ggplot(EPAair, aes(x = Ozone, color = as.factor(Year))) +
  geom_density(fill = "red", alpha = 0.1) +
  geom_vline(xintercept = 50, color = "#238b45", lty = 2, size = 0.9) + 
  scale_x_continuous(expand = c(0, 0)) + scale_y_continuous(expand = c(0, 0)) 

print(Geomdens.plot)

#2
Geomboxplot <- ggplot(EPAair, aes(x = Ozone, color = as.factor(Year))) +
  geom_density(fill = "grey") +
  geom_vline(xintercept = 50, color = "#238b45", lty = 2, size = 0.9) + 
  scale_color_manual( +
  stat_summary(x = , y = )

print(Geomboxplot)

#3
Geomviolinplot <- ggplot(EPAair, aes(x = Ozone, color = as.factor(Year))) +
  geom_density(fill = "grey") +
  geom_vline(xintercept = 50, color = "#238b45", lty = 2, size = 0.9) + 
  scale_x_continuous(expand = c(0, 0)) + scale_y_continuous(expand = c(0, 0)) +
  stat_summary(x = , y = )


```

Now, write a summary of your findings, incorporating statistical output, reference to the figure(s), and a contextual interpretation.

>
